Added tests covering the checkbox usage with setValue
