+++ driver-testsuite/tests/Form/GeneralTest.php
@@ -232,34 +232,6 @@ OUT;
         }
     }
 
-    public function testCheckboxMultiple()
-    {
-        $this->getSession()->visit($this->pathTo('/multicheckbox_form.html'));
-
-        $page = $this->getSession()->getPage();
-        $this->assertEquals('Multicheckbox Test', $page->find('css', 'h1')->getText());
-
-        $updateMail  = $page->find('css', '[name="mail_types[]"][value="update"]');
-        $spamMail    = $page->find('css', '[name="mail_types[]"][value="spam"]');
-
-        $this->assertNotNull($updateMail);
-        $this->assertNotNull($spamMail);
-
-        $this->assertTrue($updateMail->getValue());
-        $this->assertFalse($spamMail->getValue());
-
-        $this->assertTrue($updateMail->isChecked());
-        $this->assertFalse($spamMail->isChecked());
-
-        $updateMail->uncheck();
-        $this->assertFalse($updateMail->isChecked());
-        $this->assertFalse($spamMail->isChecked());
-
-        $spamMail->check();
-        $this->assertFalse($updateMail->isChecked());
-        $this->assertTrue($spamMail->isChecked());
-    }
-
     public function testMultiInput()
     {
         $this->getSession()->visit($this->pathTo('/multi_input_form.html'));
